+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/MapboxMap.kt
@@ -28,6 +28,8 @@ import com.mapbox.maps.extension.compose.ornaments.attribution.MapAttributionSco
 import com.mapbox.maps.extension.compose.ornaments.compass.MapCompassScope
 import com.mapbox.maps.extension.compose.ornaments.logo.MapLogoScope
 import com.mapbox.maps.extension.compose.ornaments.scalebar.MapScaleBarScope
+import com.mapbox.maps.extension.compose.style.MapboxStandardStyle
+import com.mapbox.maps.extension.compose.style.MapboxStyleComposable
 import com.mapbox.maps.plugin.Plugin
 import com.mapbox.maps.plugin.Plugin.Companion.MAPBOX_ATTRIBUTION_PLUGIN_ID
 import com.mapbox.maps.plugin.Plugin.Companion.MAPBOX_COMPASS_PLUGIN_ID
@@ -54,6 +56,7 @@ import kotlinx.coroutines.awaitCancellation
  * @param mapViewportState A state object that can be hoisted to control and observe the map's camera state. A [MapViewportState] may only be used by a single [MapboxMap] composable at a time as it reflects instance state for a single view of a map.
  * @param onMapClickListener Callback to be invoked when the user clicks on the map view.
  * @param onMapLongClickListener Callback to be invoked when the user long clicks on the map view.
+ * @param style The Style of the map.
  * @param content The content of the map.
  */
 @Composable
@@ -73,6 +76,7 @@ public fun MapboxMap(
   mapViewportState: MapViewportState = rememberMapViewportState(),
   onMapClickListener: OnMapClickListener = DefaultSettingsProvider.defaultOnClickListener,
   onMapLongClickListener: OnMapLongClickListener = DefaultSettingsProvider.defaultOnLongClickListener,
+  style: @Composable @MapboxStyleComposable () -> Unit = { MapboxStandardStyle() },
   content: (@Composable @MapboxMapComposable MapboxMapScope.() -> Unit)? = null
 ) {
   // display placeholder when in preview mode.
@@ -120,6 +124,7 @@ public fun MapboxMap(
   val currentOnMapLongClickListener by rememberUpdatedState(onMapLongClickListener)
   val currentContent by rememberUpdatedState(content)
   val currentMapEvents by rememberUpdatedState(mapEvents)
+  val currentStyle by rememberUpdatedState(style)
   LaunchedEffect(Unit) {
     disposingComposition(
       Composition(
@@ -135,6 +140,8 @@ public fun MapboxMap(
             currentOnMapLongClickListener,
             currentMapEvents,
           )
+          // add Style node with the styleUri
+          currentStyle.invoke()
           currentContent?.let { MapboxMapScope.it() }
         }
       }