+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/internal/MapboxMapNode.kt
@@ -186,7 +186,7 @@ private class MapboxMapNode(
       field = value
     }
 
-  override fun onAttached() {
+  override fun onAttached(parent: MapNode) {
     controller.gestures.apply {
       addNonDefaultOnClickListener(clickListener)
       addNonDefaultOnLongClickListener(longClickListener)
@@ -212,7 +212,7 @@ private class MapboxMapNode(
     }
   }
 
-  override fun onRemoved() {
+  override fun onRemoved(parent: MapNode) {
     cleanUp()
   }
 
@@ -228,6 +228,10 @@ private class MapboxMapNode(
     cancelable?.cancel()
     mapEventCancelableHolder.cancelAll()
     controller.viewport.removeStatusObserver(viewportStatusObserver)
+  }
+
+  override fun toString(): String {
+    return "MapboxMapNode()"
   }
 }