+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/style/internal/MapStyleNode.kt
@@ -1,0 +1,91 @@
+package com.mapbox.maps.extension.compose.style.internal
+
+import com.mapbox.maps.MapboxMap
+import com.mapbox.maps.StyleDataLoaded
+import com.mapbox.maps.StyleDataLoadedType
+import com.mapbox.maps.coroutine.styleDataLoadedEvents
+import com.mapbox.maps.extension.compose.internal.MapNode
+import com.mapbox.maps.logD
+import com.mapbox.maps.logW
+import kotlinx.coroutines.CoroutineName
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.cancel
+import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.flow.SharingStarted
+import kotlinx.coroutines.flow.filter
+import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.flow.shareIn
+
+internal class MapStyleNode(
+  val style: String,
+  val mapboxMap: MapboxMap,
+) : MapNode() {
+
+  val coroutineScope =
+    CoroutineScope(Dispatchers.Main.immediate + SupervisorJob() + CoroutineName("MapStyleNodeScope"))
+
+  private val styleDataLoadedFlow = mapboxMap.styleDataLoadedEvents
+
+  internal val styleSourcesLoaded: SharedFlow<StyleDataLoaded> = styleDataLoadedFlow
+    .filter { it.type == StyleDataLoadedType.SOURCES }
+    .onEach { logD(TAG, "$this - styleSourcesLoaded: $it") }
+    .shareIn(
+      scope = coroutineScope,
+      started = SharingStarted.Eagerly,
+      replay = 1
+    )
+  internal val styleDataLoaded = styleDataLoadedFlow
+    .filter { it.type == StyleDataLoadedType.STYLE }
+    .onEach { logD(TAG, "$this - styleDataLoaded: $it") }
+    .shareIn(
+      scope = coroutineScope,
+      started = SharingStarted.Eagerly,
+      replay = 1
+    )
+  internal val styleSpritesLoaded = styleDataLoadedFlow
+    .filter { it.type == StyleDataLoadedType.SPRITE }
+    .onEach { logD(TAG, "$this - styleSpritesLoaded: $it") }
+    .shareIn(
+      scope = coroutineScope,
+      started = SharingStarted.Eagerly,
+      replay = 1
+    )
+
+  override fun onAttached(parent: MapNode) {
+    logD(TAG, "onAttached: parent=$parent")
+    updateStyle(style)
+  }
+
+  override fun onRemoved(parent: MapNode) {
+    logD(TAG, "onRemoved: parent=$parent")
+    coroutineScope.cancel()
+    // remove the current style with loading an empty style json, if this is the only style
+    // node in the node tree.
+    val styleInstances = parent.children.filterIsInstance<MapStyleNode>()
+    if (styleInstances.size == 1) {
+      updateStyle("{}")
+    } else {
+      logW(TAG, "Multiple style node detected in the tree:")
+      styleInstances.forEach {
+        logW(TAG, "\t${it.style}")
+      }
+    }
+  }
+
+  private fun updateStyle(style: String) {
+    logD(TAG, "loadStyle $style started")
+    mapboxMap.loadStyle(style) {
+      logD(TAG, "loadStyle $style finished")
+    }
+  }
+
+  override fun toString(): String {
+    return "MapStyleNode(style=$style)"
+  }
+
+  private companion object {
+    private const val TAG = "MapStyleNode"
+  }
+}