+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Mapbox welcomes participation and contributions from everyone.
 * Add `SdkInformationQuery` to expose sdk version information.
 * Enable `TileStore` delta updates by default for Maps domain. 
 * Add `TileStore.estimateTileRegion` API for estimating Tile Region downloads and storage size.
+* [compose] Introduce experimental `MapStyle`, `MapboxStandardStyle`, `GenericStyle` composable functions to work with the map style.
+* [compose] Added experimental layer and source composable functions to insert layer/sources to the map.
 
 ## Bug fixes 🐞
 * [compose] Fix `ViewAnnotation` not cleared when it leaves composition.